Complete namespace changes. Add Trevor's new weighted extents and other updates. Ready for merge into trunk.

+++ indri/branches/dmf-devel/include/indri/ContextCountAccumulator.hpp
@@ -44,17 +44,17 @@ namespace indri
       ListIteratorNode* _matches; 
 
       std::string _name;
-      UINT64 _occurrences;
-      UINT64 _contextSize;
+  double _occurrences;
+  double _contextSize;
 
       EvaluatorNode::MResults _results;
 
-    public:
+public:
       ContextCountAccumulator( const std::string& name, ListIteratorNode* matches, ListIteratorNode* context );
       ~ContextCountAccumulator();
 
-      UINT64 getOccurrences() const;
-      UINT64 getContextSize() const;
+  double getOccurrences() const;
+  double getContextSize() const;
 
       const ListIteratorNode* getContextNode() const;
       const ListIteratorNode* getMatchesNode() const;
